rename calculate_raster.py to grid_to_raster
