+++ hhnk_threedi_tools/core/checks/one_d_two_d.py
@@ -8,7 +8,7 @@ from shapely.geometry import LineString
 
 import hhnk_threedi_tools.core.checks.grid_result_metadata as grid_result_metadata
 from hhnk_threedi_tools.core.folders import Folders
-from hhnk_threedi_tools.core.result_rasters.calculate_raster import GridToRaster
+from hhnk_threedi_tools.core.result_rasters.grid_to_raster import GridToRaster
 
 # Local imports
 from hhnk_threedi_tools.core.result_rasters.netcdf_to_gridgpkg import NetcdfToGPKG
